Expose API command to start a func task and to check if CLI tools is installed (#4299)
* Add an API command to start a func task and to check if CLI tools is installed

* Make some updates for isFuncCoreToolsInstalled

* Update apiVersion

* Update apiVersion to minor path

* Use fully qualified path rather than relative

* Use Uri.fspath as key for FuncRunningMap/Port (#4309)

* Use Uri.fspath as key for FuncRunningMap/Port

This commit modifies pickFuncProcess startFuncProcessFromApi removing
the workspaceFolder requirement and adding in the environment variables.

funcHostTask was originally tracked by vscode.WorkspaceFolder via
Task.Scope and DebugConfiguration.WorkspaceFolder. This commit modifies
it to use a new AzureFunctionTaskDefinition.functionsApp property which
is either the `buildPath` from startFuncProcessFromApi or the
vscode.WorkspaceFolder.uri.fsPath to maintain backwards compatibility.

runningFuncTaskMap, funcTaskStartedEmitter, and runningFuncPortMap now
all using the vscode.Uri.fsPath string which can be used to map back to
a WorkspaceFolder. However in the cases the WorkspaceFolder fails, APIs
that used it will now see undefined passed in.

* Make unique 'type' so VS Code will launch multiple tasks

* Address Comments

* Pin core tools to fix tests (#4310)

* Try stable v4 feed

* Fixup

---------

+++ src/funcCoreTools/funcHostTask.ts
@@ -7,20 +7,43 @@ import { registerEvent, type IActionContext } from '@microsoft/vscode-azext-util
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { tryGetFunctionProjectRoot } from '../commands/createNewProject/verifyIsProject';
-import { localSettingsFileName } from '../constants';
+import { func, localSettingsFileName } from '../constants';
 import { getLocalSettingsJson } from '../funcConfig/local.settings';
 import { getWorkspaceSetting } from '../vsCodeConfig/settings';
 
 export interface IRunningFuncTask {
+    taskExecution: vscode.TaskExecution;
     processId: number;
 }
 
-export const runningFuncTaskMap: Map<vscode.WorkspaceFolder | vscode.TaskScope, IRunningFuncTask> = new Map<vscode.WorkspaceFolder | vscode.TaskScope, IRunningFuncTask>();
+export class AzureFunctionTaskDefinition implements vscode.TaskDefinition {
+    type: string;
+    // This is either:
+    // - vscode.WorkspaceFolder.uri.fsPath (used for most of the scenarios in this extension.)
+    // - If using the exported API 'startFuncProcessFromApi', it will be the binary path wrapped with vscode.Uri.file().fsPath
+    functionsApp: string
 
-const funcTaskStartedEmitter = new vscode.EventEmitter<vscode.WorkspaceFolder | vscode.TaskScope | undefined>();
+    static is(taskDefinition: vscode.TaskDefinition): taskDefinition is AzureFunctionTaskDefinition {
+        return taskDefinition.type.startsWith(func) && "functionsApp" in taskDefinition;
+    }
+}
+
+interface DotnetDebugDebugConfiguration extends vscode.DebugConfiguration {
+    launchServiceData: { [key: string]: string }
+}
+
+namespace DotnetDebugDebugConfiguration {
+    export function is(debugConfiguration: vscode.DebugConfiguration): debugConfiguration is DotnetDebugDebugConfiguration {
+        return debugConfiguration.type === 'coreclr' && 'launchServiceData' in debugConfiguration
+    }
+}
+
+export const runningFuncTaskMap: Map<string, IRunningFuncTask> = new Map<string, IRunningFuncTask>();
+
+const funcTaskStartedEmitter = new vscode.EventEmitter<string>();
 export const onFuncTaskStarted = funcTaskStartedEmitter.event;
 
-export const runningFuncPortMap = new Map<vscode.WorkspaceFolder | vscode.TaskScope | undefined, string>();
+export const runningFuncPortMap = new Map<string | undefined, string>();
 const defaultFuncPort: string = '7071';
 
 export function isFuncHostTask(task: vscode.Task): boolean {
@@ -32,18 +55,20 @@ export function registerFuncHostTaskEvents(): void {
     registerEvent('azureFunctions.onDidStartTask', vscode.tasks.onDidStartTaskProcess, async (context: IActionContext, e: vscode.TaskProcessStartEvent) => {
         context.errorHandling.suppressDisplay = true;
         context.telemetry.suppressIfSuccessful = true;
-        if (e.execution.task.scope !== undefined && isFuncHostTask(e.execution.task)) {
-            runningFuncTaskMap.set(e.execution.task.scope, { processId: e.processId });
-            runningFuncPortMap.set(e.execution.task.scope, await getFuncPortFromTaskOrProject(context, e.execution.task, e.execution.task.scope));
-            funcTaskStartedEmitter.fire(e.execution.task.scope);
+        if (AzureFunctionTaskDefinition.is(e.execution.task.definition) && isFuncHostTask(e.execution.task)) {
+            const workspaceFolder: vscode.WorkspaceFolder | undefined = vscode.workspace.getWorkspaceFolder(vscode.Uri.parse(e.execution.task.definition.functionsApp))
+
+            runningFuncTaskMap.set(e.execution.task.definition.functionsApp, { taskExecution: e.execution, processId: e.processId });
+            runningFuncPortMap.set(e.execution.task.definition.functionsApp, await getFuncPortFromTaskOrProject(context, e.execution.task, workspaceFolder));
+            funcTaskStartedEmitter.fire(e.execution.task.definition.functionsApp);
         }
     });
 
     registerEvent('azureFunctions.onDidEndTask', vscode.tasks.onDidEndTaskProcess, (context: IActionContext, e: vscode.TaskProcessEndEvent) => {
         context.errorHandling.suppressDisplay = true;
         context.telemetry.suppressIfSuccessful = true;
-        if (e.execution.task.scope !== undefined && isFuncHostTask(e.execution.task)) {
-            runningFuncTaskMap.delete(e.execution.task.scope);
+        if (AzureFunctionTaskDefinition.is(e.execution.task.definition) && isFuncHostTask(e.execution.task)) {
+            runningFuncTaskMap.delete(e.execution.task.definition.functionsApp);
         }
     });
 
@@ -51,28 +76,41 @@ export function registerFuncHostTaskEvents(): void {
         context.errorHandling.suppressDisplay = true;
         context.telemetry.suppressIfSuccessful = true;
 
+        // Used to stop the task started with pickFuncProcess.ts startFuncProcessFromApi.
+        if (DotnetDebugDebugConfiguration.is(debugSession.configuration) && debugSession.configuration.launchServiceData.buildPath) {
+            const buildPathUri: vscode.Uri = vscode.Uri.file(debugSession.configuration.launchServiceData.buildPath)
+            stopFuncTaskIfRunning(buildPathUri.fsPath, /* terminate */ true)
+        }
+
         // NOTE: Only stop the func task if this is the root debug session (aka does not have a parentSession) to fix https://github.com/microsoft/vscode-azurefunctions/issues/2925
         if (getWorkspaceSetting<boolean>('stopFuncTaskPostDebug') && !debugSession.parentSession && debugSession.workspaceFolder) {
-            stopFuncTaskIfRunning(debugSession.workspaceFolder);
+            stopFuncTaskIfRunning(debugSession.workspaceFolder.uri.fsPath);
         }
     });
 }
 
-export function stopFuncTaskIfRunning(workspaceFolder: vscode.WorkspaceFolder): void {
-    const runningFuncTask: IRunningFuncTask | undefined = runningFuncTaskMap.get(workspaceFolder);
+export function stopFuncTaskIfRunning(functionApp: string, terminate: boolean = false): void {
+    const runningFuncTask: IRunningFuncTask | undefined = runningFuncTaskMap.get(functionApp);
     if (runningFuncTask !== undefined) {
-        // Use `process.kill` because `TaskExecution.terminate` closes the terminal pane and erases all output
-        // Also to hopefully fix https://github.com/microsoft/vscode-azurefunctions/issues/1401
-        process.kill(runningFuncTask.processId);
-        runningFuncTaskMap.delete(workspaceFolder);
+        if (terminate) {
+            // Tasks that are spun up by an API will execute quickly that process.kill does not terminate the func host fast enough
+            // that it hangs on to the port needed by a debug-restart event.
+            runningFuncTask.taskExecution.terminate();
+        }
+        else {
+            // Use `process.kill` because `TaskExecution.terminate` closes the terminal pane and erases all output
+            // Also to hopefully fix https://github.com/microsoft/vscode-azurefunctions/issues/1401
+            process.kill(runningFuncTask.processId);
+        }
+        runningFuncTaskMap.delete(functionApp);
     }
 }
 
-export async function getFuncPortFromTaskOrProject(context: IActionContext, funcTask: vscode.Task | undefined, projectPathOrTaskScope: string | vscode.WorkspaceFolder | vscode.TaskScope): Promise<string> {
+export async function getFuncPortFromTaskOrProject(context: IActionContext, funcTask: vscode.Task | undefined, projectPathOrTaskScope: string | vscode.WorkspaceFolder | vscode.TaskScope | undefined): Promise<string> {
     try {
         // First, check the task itself
-        if (funcTask && typeof funcTask.definition.command === 'string') {
-            const match = funcTask.definition.command.match(/\s+(?:"|'|)(?:-p|--port)(?:"|'|)\s+(?:"|'|)([0-9]+)/i);
+        if (funcTask && funcTask.execution instanceof vscode.ShellExecution) {
+            const match = funcTask.execution?.commandLine?.match(/\s+(?:"|'|)(?:-p|--port)(?:"|'|)\s+(?:"|'|)([0-9]+)/i);
             if (match) {
                 return match[1];
             }
